Add an empty maps for each robot kind.
Empty maps are useful to test the robot's control without any obstacles.

Updated the links in .world files to more recently version of the stage simulator. Before were pointing to Stage v2.1.0 and now they are pointing to Stage v3.2.1.

Changed the color of the holonomic robot base, because in the stage simulator a red footprint is easier to see than a grey one.